Some custom stuff with the Photo Upload funtionality

+++ Models/Order.cs
@@ -18,7 +18,7 @@ namespace GeorgianGroceries.Models
         public string PostalCode { get; set; }
         public string Phone { get; set; }
         public string email { get; set; }
-        public float Total { get; set; }
+        public double Total { get; set; }
         //Child reference 
         public List<OrderDetail> OrderDetails { get; set; }
     }
